iot-service/pyproject.toml
[project]
name = "iot-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Tetsu",email = "tetsufromtw@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
